+++ CHANGELOG.md
@@ -10,6 +10,7 @@ __! Be certain to carefully back up any critical user data before installing thi
 - Added `itemUuid` to Roll/RnK for technique and weapons to be readable in ChatMessage (use `fromUuid()` / `fromUuidSync()` to get the object).
 - Added wiki link in system tab.
 - Removed restriction on technique types when dropping a technique (Sheet and 20Q. #39).
+- Fixed sync between GM for Combat tracker `initiative encounter type` and `initiative prepared`.
 
 ### OpenDicePicker API usage
 #### Fitness skill roll for the all combat targets